pyproject.toml
[project]
name = "project3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.3.1",
    "fastapi>=0.122.0",
    "httpx>=0.28.1",
    "ipynb>=0.5.1",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-experimental>=0.4.0",
    "langchain-google-community>=3.0.1",
    "langchain-groq>=1.1.0",
    "langchain-mcp-adapters>=0.1.14",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "langgraph-cli[inmem]>=0.4.7",
    "langsmith>=0.4.49",
    "mcp>=1.22.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "tavily-python>=0.7.13",
    "uvicorn>=0.38.0",
    "wikipedia>=1.4.0",
]

[tool.uv.workspace]
members = [
    ".",
    ".",
]

[tool.uv.sources]
project3 = { workspace = true }

[dependency-groups]
dev = [
    "arxiv>=2.3.1",
    "fastapi>=0.122.0",
    "httpx>=0.28.1",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-experimental>=0.4.0",
    "langchain-google-community>=3.0.1",
    "langchain-google-genai>=3.2.0",
    "langchain-groq>=1.1.0",
    "langchain-mcp-adapters>=0.1.14",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "langgraph-cli[inmem]>=0.4.7",
    "langsmith>=0.4.49",
    "mcp>=1.22.0",
    "project3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
    "tavily-python>=0.7.13",
    "uvicorn>=0.38.0",
    "wikipedia>=1.4.0",
]
